% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWDMr_DDHO.R
\docType{methods}
\name{AddEffect}
\alias{AddEffect}
\title{Add Additive factor to the model}
\usage{
AddEffect(object, AddName, value = numeric(0))
}
\arguments{
\item{object}{An SWDMr_DDHO object}

\item{AddName}{character present in SWdist data.frame}

\item{value}{if this is a free parameter, let it be empty.}
}
\description{
Add Additive factor to the model
}
\examples{
model<-AddEffect(model,"SD")
}
