% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{AddForce}
\alias{AddForce}
\title{Add Force to the model}
\usage{
AddForce(object, ForceName, value = numeric(0))
}
\arguments{
\item{object}{An SWDMr_DDHO object}

\item{ForceName}{character present in SWdist data.frame}

\item{value}{if this is a free parameter, let it be empty.}
}
\description{
Add Force to the model
}
\examples{
model<-AddForce(model,"Wake")
}
