% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWDMr_DDHO.R
\docType{methods}
\name{AddSinF}
\alias{AddSinF}
\title{Add Sin Force to the model}
\usage{
AddSinF(object, FixAmp = numeric(0), FixPhi = numeric(0),
  FixPer = numeric(0))
}
\arguments{
\item{object}{An SWDMr_DDHO object}

\item{FixAmp}{Fix Sin Force Amplitude to the given value}

\item{FixPhi}{Fix the Sin Force phase to the given value}

\item{FixPer}{Fix the period (in hour) of the Sin Force}
}
\description{
Add Sin Force to the model
}
\examples{
model<-AddSinF(model,FixPer = 24)
}
