% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{FixAsymptSleep}
\alias{FixAsymptSleep}
\title{Fix Asymptote of sleep in the model}
\usage{
FixAsymptSleep(object, value)
}
\arguments{
\item{object}{An SWDMr_ProcS object}

\item{value}{top value reached by wake}
}
\description{
Fix Asymptote of sleep in the model
}
\examples{
model<-FixAsymptSleep(model,80)
}
