% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{FixAsymptWake}
\alias{FixAsymptWake}
\title{Fix Asymptote of wake in the model}
\usage{
FixAsymptWake(object, value)
}
\arguments{
\item{object}{An SWDMr_ProcS object}

\item{value}{top value reached by wake}
}
\description{
Fix Asymptote of wake in the model
}
\examples{
model<-FixAsymptWake(model,300)
}
