% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{FixDamping}
\alias{FixDamping}
\title{Fix damping of the model}
\usage{
FixDamping(object, value)
}
\arguments{
\item{object}{An SWDMr_DDHO object}

\item{value}{log gamma}
}
\description{
Fix damping of the model
}
\examples{
model<-FixNaturalFrequency(model,-2)
}
