% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{FixIntercept}
\alias{FixIntercept}
\title{Fix intercept of the model}
\usage{
FixIntercept(object, value)
}
\arguments{
\item{object}{An SWDMr_DDHO object}

\item{value}{intercept}
}
\description{
Fix intercept of the model
}
\examples{
model<-FixIntercept(model,0)
}
