% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{FixNaturalFrequency}
\alias{FixNaturalFrequency}
\title{Fix natural frequency of the model}
\usage{
FixNaturalFrequency(object, value)
}
\arguments{
\item{object}{An SWDMr_DDHO object}

\item{value}{natural frequency}
}
\description{
Fix natural frequency of the model
}
\examples{
model<-FixNaturalFrequency(model,2*pi/24)
}
