% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{FixTimeConstSleep}
\alias{FixTimeConstSleep}
\title{Fix Time constant of sleep in the model}
\usage{
FixTimeConstSleep(object, value)
}
\arguments{
\item{object}{An SWDMr_ProcS object}

\item{value}{top value reached by wake}
}
\description{
Fix Time constant of sleep in the model
}
\examples{
model<-FixTimeConstSleep(model,8)
}
