% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{FixTimeConstWake}
\alias{FixTimeConstWake}
\title{Fix time constant of wake in the model}
\usage{
FixTimeConstWake(object, value)
}
\arguments{
\item{object}{An SWDMr_ProcS object}

\item{value}{top value reached by wake}
}
\description{
Fix time constant of wake in the model
}
\examples{
model<-FixTimeConstWake(model,16)
}
