% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{GetFit}
\alias{GetFit}
\title{GetFit data given model}
\usage{
GetFit(object, params)
}
\arguments{
\item{object}{An SWDMr model object}

\item{params}{a named vector with required free parameter value}
}
\description{
GetFit data given model
}
\examples{
fitted<-GetFit(model)
}
