% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\name{initProcessSmodel}
\alias{initProcessSmodel}
\title{Initiate Process S model}
\usage{
initProcessSmodel(object, VarExp)
}
\arguments{
\item{object}{A SWDMr object}

\item{VarExp}{The name of the variable to explain by the model}
}
\value{
SWDMr model object
}
\description{
Initiate a Process S model of sleep
}
\examples{
model <- initProcessSmodel(swdmr,VarExp = "Arntl")

}
