% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWDMr_DDHO.R
\docType{methods}
\name{PenalizeUnstableFit}
\alias{PenalizeUnstableFit}
\title{Unstable Fit penalization}
\usage{
PenalizeUnstableFit(object, value = T, PredictedValueInterval,
  StabilityDayCheck)
}
\arguments{
\item{object}{An SWDMr_DDHO object}

\item{value}{Set the penalization method (TRUE or FALSE)}

\item{PredictedValueInterval}{A two value vector for min and max value observed}

\item{StabilityDayCheck}{The number of day control for similiar min and max value compared to PredictedValueInterval}
}
\description{
Penalize the fit that are unstable through replicated points
}
\examples{
# Example of 20 day of baseline replicated
model<-ReplicateDrivingForce(model,c(0,24),20)
# Then we control that the last 10 replicated day have similar min and max value compared to true baseline
model<-PenalizeUnstableFit(model,value = T,PredictedValueInterval = c(0,24), StabilityDayCheck = 10)
}
