% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SleepDataManipulation.R
\name{Read_SW}
\alias{Read_SW}
\title{Create a dataframe with Sleep-wake distribution, Time and Day}
\usage{
Read_SW(files, concat = "mean", nepochs = 86400, epochlengthsec = 4,
  concattimesec = 300)
}
\arguments{
\item{concat}{character [mean or median]}

\item{nepochs}{integer}

\item{epochlengthsec}{integer}

\item{concattimesec}{integer}

\item{file}{character vector}
}
\value{
dataframe
}
\description{
Details in description
}
\examples{
files <- list.files(path="Data/",pattern = paste("^","BL6",sep=""),full.names = T)
SWdf<-Read_SW(files)

}
