% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWDMr.R
\docType{methods}
\name{ReplicateDrivingForce}
\alias{ReplicateDrivingForce}
\title{Replicate baseline driving force}
\usage{
ReplicateDrivingForce(object, interval, Nrep)
}
\arguments{
\item{object}{An SWDMr_DDHO object}

\item{interval}{A two value time interval that is a 24h interval (time given is included, i.e. <= or >=)}

\item{Nrep}{Number of repliated time interval}
}
\description{
Replicate the baseline (given time interval) and happend it at the begining of your Force dataset
}
\examples{
swdmr<-ReplicateDrivingForce(swdmr,c(0,24),10)
}
