% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_RNAexpression.R
\docType{data}
\name{RNAExpression}
\alias{RNAExpression}
\title{RNA-seq Time course}
\format{
A data.frame.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE140345}{Gene Expression Omnibus}
}
\usage{
data(RNAExpression)
}
\description{
Data for 4 genes (Homer1, Arntl, Acot11 & Cyth3) expression in the
mouse cortex, in a time-course experiment.
}
\details{
Sleep deprivation was applied between Time48 and Time54
}
\examples{
data(RNAExpression)
}
\references{
C.N. Hor et al. (2019) PNAS 116 (51), 25773-25783
(\href{https://pubmed.ncbi.nlm.nih.gov/31776259/}{PubMed})
}
\keyword{datasets}
