% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{SetFittingValue}
\alias{SetFittingValue}
\title{Fitting value}
\usage{
SetFittingValue(object, value = "RSS")
}
\arguments{
\item{object}{An SWDMr_DDHO object}

\item{value}{Set the fitting value return, can be Residual Sum of Square (RSS), Log likelihood (LL) or Negative Log Likelihood (NLL)}
}
\description{
Fitting value
}
\examples{
model<-SetFittingValue(model,value = "RSS")
}
