% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{SetParametersModel}
\alias{SetParametersModel}
\title{Set parameters of the model}
\usage{
SetParametersModel(object)
}
\arguments{
\item{object}{An SWDMr_DDHO object}
}
\description{
Set parameters of the model
}
\examples{
model<-SetParametersModel(model)
}
