% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{SetYinitMode}
\alias{SetYinitMode}
\title{Set the start position and speed of the oscillator}
\usage{
SetYinitMode(object, mode = "Free", values = numeric(0))
}
\arguments{
\item{object}{An SWDMr_DDHO object}

\item{mode}{character can be Free (default), Fixed, Intercept_0 or CircadianFit}

\item{values}{if the mode is "Fixed", the values is start and speed of the oscillator (e.g. c(0,0) for start at 0 and speed of 0. If the mode is CircadianFit, values are the time interval for the circadian fit (e.g. c(0,48) for time 0 to time 48)}
}
\description{
Set the start position and speed of the oscillator
}
\examples{
model<-SetYinitMode(model,"Intercept_0")
}
