% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_SleepWakeData.R
\docType{data}
\name{SleepWakeData}
\alias{SleepWakeData}
\title{C57BL/6J vigilant states}
\format{
A matrix of character
}
\source{
\href{https://doi.org/10.6084/m9.figshare.c.4421327.v1}{DOI}
}
\usage{
data(SleepWakeData)
}
\description{
Vigilant states of 12 C57BL/6J mice:
}
\details{
Wake is defined as "w" or "1"

NREM sleep is defined as "n" or "2"

REM sleep is defined as "r" or "3"

vigilant states are epochs of 4 seconds
}
\examples{
data(SleepWakeData)
}
\references{
Diessler et al. (2018) PLOS Biology 16:e2005750
(\href{https://pubmed.ncbi.nlm.nih.gov/30091978/}{PubMed})

Jan et al. (2019) Scientific Data, 6 (1) (2019), p. 258
(\href{https://pubmed.ncbi.nlm.nih.gov/31672980/}{Pubmed})
}
\keyword{datasets}
