% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_SWdf.R
\docType{data}
\name{SWdf}
\alias{SWdf}
\title{C57BL/6J Sleep-Wake amount}
\format{
A data.frame.
}
\usage{
data(SWdf)
}
\description{
Mean sleep-wake amount of 12 C57BL/6J mice.
}
\details{
Time in [hours]
}
\examples{
data(SWdf)
}
\references{
Diessler et al. (2018) PLOS Biology 16:e2005750
(\href{https://pubmed.ncbi.nlm.nih.gov/30091978/}{PubMed})

Jan et al. (2019) Scientific Data, 6 (1) (2019), p. 258
(\href{https://pubmed.ncbi.nlm.nih.gov/31672980/}{Pubmed})
}
\keyword{datasets}
