% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SleepDataManipulation.R
\name{SWdf_AddLD}
\alias{SWdf_AddLD}
\title{Add a light-dark cycle, starting at 12}
\usage{
SWdf_AddLD(SWdf)
}
\arguments{
\item{SWdf}{dataframe generated by Read_SW}
}
\value{
dataframe
}
\description{
Details in description
}
\examples{
SWdf<-AddLD(SWdf)

}
