% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SleepDataManipulation.R
\name{SWdf_DayMerging}
\alias{SWdf_DayMerging}
\title{Merging some days defined in sleep-wake table}
\usage{
SWdf_DayMerging(SWdf, Daysformat, concattimesec = 300, TimeFormatsec = 3600)
}
\arguments{
\item{SWdf}{dataframe generated by Read_SW}

\item{Daysformat}{vector for merging days}

\item{concattimesec}{integer, time frame in seconds}

\item{TimeFormatsec}{integer, time scale to 1 in seconds (3600 for 1 hour)}
}
\value{
dataframe
}
\description{
Details in description
}
\examples{
SWdf<-SWDayMerging(SWdf,Daysformat=list(c(1,2),3,4,c(1,2)))

}
