% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\name{SWDMr}
\alias{SWDMr}
\title{Create a SWDMr object}
\usage{
SWDMr(SWdist, Gexp, verbose = 1)
}
\arguments{
\item{SWdist}{sleep-wake dataframe, must contain the column: Wake, Sleep and Time}

\item{Gexp}{Gene expression dataframe which must contain genes as column and a Time column corresponding to the sleep - wake dataframe}
}
\value{
SWDMr object
}
\description{
Create a SWDMr object using a sleep-wake data.frame and a gene expression data.frame
}
\examples{
swdmr <- SWDMr(SWdist=SWdf, Gexp=Gexpdf)

}
