% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{SWDMrGetEvalFun}
\alias{SWDMrGetEvalFun}
\title{Return function to evaluate}
\usage{
SWDMrGetEvalFun(object)
}
\arguments{
\item{object}{An SWDMr model object}
}
\description{
Return function to evaluate
}
\examples{
objfun<-SWDMrGetEvalFun(model)
optimx(objfunc,params)
}
