% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{SWDMrStats}
\alias{SWDMrStats}
\title{Return statistics of the fitting}
\usage{
SWDMrStats(object, fitted, FittingValue = "RSS", detailed = F)
}
\arguments{
\item{object}{An SWDMr model object}

\item{fitted}{fitted value}

\item{FittingValue}{What value to return for the fitting (RSS or NLL or LL)}

\item{detailed}{If set to True, return all value (RSS,NLL,LL,BIC), N, nparams}
}
\description{
Return statistics of the fitting
}
\examples{
fitted<-SWDMrFit(model,params)
stats<-SWDMrStats(model,fitted)
}
