% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\name{SWdrivenValue}
\alias{SWdrivenValue}
\title{Sleep-wake driven value}
\usage{
SWdrivenValue(object, params, interval)
}
\arguments{
\item{object}{A SWDMr model}

\item{params}{parameter fitted by optimx}

\item{interval}{Time interval for the computation}
}
\value{
numeric value
}
\description{
Function to compute the Sleep-Wake driven value. As variance from the force applied 
by sleep & wake over the variance from other external forces
}
\examples{
SWdrivenValue(model,fits,c(0,24))

}
