% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{SWDMrFit}
\alias{SWDMrFit}
\title{Fit data given model}
\usage{
SWDMrFit(object, params)
}
\arguments{
\item{object}{An SWDMr model object}

\item{params}{a named vector with required free parameter value}
}
\description{
Fit data given model
}
\examples{
fitted<-SWDMrFit(model)
}
