% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\docType{methods}
\name{SWDMrGetEvalFun}
\alias{SWDMrGetEvalFun}
\title{Return function to evaluate}
\usage{
SWDMrGetEvalFun(object, match = "exact", method = "rk4")
}
\arguments{
\item{object}{An SWDMr model object}

\item{match}{character "exact" or "approx"}

\item{method}{method to solve ODE, approximation by "rk4" or true "Solve"}
}
\description{
Return function to evaluate
}
\examples{
objfun<-SWDMrGetEvalFun(model)
optimx(objfunc,params)
}
